Fixed some size related warnings by using NSInteger (which should always be the same size as a pointer).
I think the code is a bit dubious though - not sure why you can't store a boxed bool value in the cache instead. That should returns NULL if it's not been set, or an NSValue object if it has (which you can then convert into a BOOL).

+++ src/classes/GRMustacheContext.m
@@ -677,7 +677,7 @@ NSString *canonicalKeyForKey(Class klass, NSString *key);
     });
     
     Class klass = object_getClass(object);
-    int result = (int)CFDictionaryGetValue(cache, klass);   // 0 = undefined, 1 = YES, 2 = NO
+    NSInteger result = (NSInteger)CFDictionaryGetValue(cache, klass);   // 0 = undefined, 1 = YES, 2 = NO
     if (!result) {
         Class NSOrderedSetClass = NSClassFromString(@"NSOrderedSet");
         result = ([klass isSubclassOfClass:[NSArray class]] ||
@@ -727,7 +727,7 @@ NSString *canonicalKeyForKey(Class klass, NSString *key);
     });
     
     Class klass = object_getClass(object);
-    int result = (int)CFDictionaryGetValue(cache, klass);   // 0 = undefined, 1 = YES, 2 = NO
+    NSInteger result = (int)CFDictionaryGetValue(cache, klass);   // 0 = undefined, 1 = YES, 2 = NO
     if (!result) {
         result = class_conformsToProtocol(klass, @protocol(GRMustacheTagDelegate)) ? 1 : 2;
         CFDictionarySetValue(cache, klass, (const void *)result);
